refactor(FundStatus): Change format of FundStatus Enum
