+++ src/main/scala/com/particeep/api/models/enums/FundStatus.scala
@@ -1,18 +1,14 @@
 package com.particeep.api.models.enums
 
-object FundStatus {
+sealed trait FundStatus extends Product with Enum
 
-  sealed abstract class FundStatus extends Product with Enum
-
+object FundStatus extends EnumHelper[FundStatus] {
   case object INIT         extends FundStatus
   case object UNDER_REVIEW extends FundStatus
   case object RUNNING      extends FundStatus
   case object SUCCEEDED    extends FundStatus
 
-  object FundStatus extends EnumHelper[FundStatus] {
-    def values: Set[FundStatus] = Set(INIT, UNDER_REVIEW, RUNNING, SUCCEEDED)
+  def values: Set[FundStatus] = Set(INIT, UNDER_REVIEW, RUNNING, SUCCEEDED)
 
-    implicit def stringToFundStatus(value: String): FundStatus = get(value.toUpperCase).getOrElse(INIT)
-  }
-
+  implicit def stringToFundStatus(value: String): FundStatus = get(value.toUpperCase).getOrElse(INIT)
 }
