+++ src/main/scala/com/particeep/api/models/fund/Fund.scala
@@ -8,9 +8,9 @@ import java.time.OffsetDateTime
 
 import com.particeep.api.core.Formatter
 import com.particeep.api.models.Address
-import com.particeep.api.models.enums.Currency
 import com.particeep.api.models.enums.Currency.EUR
-import com.particeep.api.models.enums.FundStatus.{ FundStatus, INIT }
+import com.particeep.api.models.enums.FundStatus.INIT
+import com.particeep.api.models.enums.{ Currency, FundStatus }
 
 case class Fund(
   id:                           String                 = "",
